Test to verify integer operation on foreign objects
